+++ src/growthbook.ts
@@ -1,6 +1,10 @@
-import { GrowthBookClient, setPolyfills } from "@growthbook/growthbook"
+import { GrowthBookClient, setPolyfills, type UserScopedGrowthBook, type Attributes } from "@growthbook/growthbook"
 import { EventSource } from "eventsource"
+import type * as express from "express"
+import { isbot } from "isbot"
+import { createContext } from "react-router"
 import logger from "./logger.js"
+import { BrowserDetection } from "./util/browser_detection.js"
 
 setPolyfills({
   EventSource,
@@ -27,3 +31,42 @@ export const createGrowthBook = async (config: GrowthBookConfig) => {
 
   return growthbook
 }
+
+export interface ScopedGrowthbookOptions {
+  attributes?: Attributes
+  forcedFeatures?: Map<string, unknown>
+  deviceId?: string
+}
+
+export const createScopedGrowthBook = (
+  request: express.Request,
+  client: GrowthBookClient,
+  options?: ScopedGrowthbookOptions
+) => {
+  const userAgent = request.headers["user-agent"] || ""
+  const bot = isbot(userAgent)
+
+  const gbInstance = client.createScopedInstance({
+    attributes: {
+      url: request.url,
+      path: request.path,
+      host: request.headers["host"],
+      deviceType: BrowserDetection.mobile(userAgent) ? "mobile" : "desktop",
+      browser: BrowserDetection.browser(userAgent),
+      bot,
+      deviceId: options?.deviceId,
+    },
+  })
+
+  if (options?.forcedFeatures) {
+    gbInstance.setForcedFeatures(options.forcedFeatures)
+  }
+
+  if (options?.attributes) {
+    gbInstance.updateAttributes(options.attributes)
+  }
+
+  return gbInstance
+}
+
+export const growthbookContext = createContext<UserScopedGrowthBook | undefined>()